# ENH/API: DataFrame.stack() support for level=None, sequentially=True/False, and NaN level values.