--- a/pandas/tools/tests/test_pivot.py
+++ b/pandas/tools/tests/test_pivot.py
@@ -431,6 +431,7 @@ class TestPivotTable(tm.TestCase):
                              columns='Carl Joe Mark'.split())
         expected.index.name = 'Date'
         expected.columns.name = 'Buyer'
+        expected['Carl'] = expected['Carl'].astype(df.dtypes['Quantity'])
 
         result = pivot_table(df, index=Grouper(freq='6MS'), columns='Buyer',
                              values='Quantity', aggfunc=np.sum)