… pandas/core/reshape.py
… pandas/tests/test_frame.py