closes #8851
closes #9399 
closes #9406
closes #9533