@seth-p 2015-01-22
Haven't looked at this in a while. Will try to revisit over next few days.