@seth-p 2015-02-02
@jreback, @shoyer, question for you guys: Is ``pd.core.reshape.stack()`` a "public" API, that I should not alter (except in a backwards compatible way), or am I free to change it as I see fit so long as ``pd.DataFrame.stack()`` remains backwards compatible? I think it should be regarded as internal -- it doesn't show up in the API Reference, http://pandas.pydata.org/pandas-docs/version/0.15.2/api.html (though that's not necessarily complete) -- and that I can change it if I want, but figured I'd check first.