@shoyer 2015-02-02
@seth-p I agree, I don't think we make any API guarantees for `pandas.core` -- only for the top level namespace.