@jorisvandenbossche 2015-02-03
I don't really like the `pd.Index.ALL_LEVELS`. Can't we use `None` for this? `df.unstack(level=None)`