@seth-p 2015-02-03
@jorisvandenbossche, ok, I've removed ``ALL_LEVELS`` and made ``df.stack(level=None)`` mean all levels.

I haven't made any corresponding changes to ``DataFrame.unstack()``, as that code is very different from ``DataFrame.stack()``. I would have expected ``DataFrame.unstack()`` to be implemented as ``DataFrame.T.stack().T``, but it's not (except that it is implemented as ``DataFrame.T.stack()`` in the special case that ``level`` is a single number and ``df.index`` is a simple ``Index`` (not a ``MultiIndex``).)

In general the code (in ``reshape.py``) that implements ``DataFrame.unstack()`` (``unstack()``, ``_unstack_frame()``, and ``_unstack_multiple()``) looks quite different from the code that implements ``DataFrame.stack()`` (``stack()``, ``stack_multiple()``, and ``_stack_multi_columns()`` -- before the changes of this PR).

Comparing the results of ``df.unstack()`` to ``df.T.stack().T``, I observe two differences:

1.  dtypes. Since dtypes are done by column, transposing can mess up dtypes. Given this,I'm still thinking what to do about ``unstack()``.

2. When a MultiIndex label is ``NaN``, things are just plain weird. Perhaps there should be a separate issue for this. While the current code passes the ``test_unstack_nan_index()`` in ``test_frame.py``, the treatment of ``NaN`` labels seems very inconsistent/unreliable:
```
In [140]: df = pd.DataFrame(np.arange(4).reshape(2, 2),
                            columns=pd.MultiIndex.from_tuples([('A','a'), ('B', 'b')],
                                                              names=['Upper', 'Lower']),
                            index=Index([0, 1], name='Num'), dtype=np.float64)

In [141]: df = pd.DataFrame(np.arange(4).reshape(2, 2),
                            columns=pd.MultiIndex.from_tuples([('A',np.nan), ('B', 'b')],
                                                              names=['Upper', 'Lower']),
                            index=Index([0, 1], name='Num'), dtype=np.float64)

In [148]: df
Out[148]:
Upper  A  B
Lower  a  b
Num
0      0  1
1      2  3

In [149]: df.stack()
Out[149]:
Upper       A   B
Num Lower
0   a       0 NaN
    b     NaN   1
1   a       2 NaN
    b     NaN   3

In [150]: df.T.unstack().T
Out[150]:
Upper       A   B
Num Lower
0   a       0 NaN
    b     NaN   1
1   a       2 NaN
    b     NaN   3

In [151]: df_nan
Out[151]:
Upper   A  B
Lower NaN  b
Num
0       0  1
1       2  3

In [152]: df_nan.stack()
Out[152]:
Upper      A  B
Num Lower
0   NaN    0  1
    b      0  1
1   NaN    2  3
    b      2  3

In [153]: df_nan.T.unstack().T
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-153-edbcaeb64f64> in <module>()
----> 1 df_nan.T.unstack().T

C:\Python34\lib\site-packages\pandas\core\frame.py in unstack(self, level)
   3486         """
   3487         from pandas.core.reshape import unstack
-> 3488         return unstack(self, level)
   3489
   3490     #----------------------------------------------------------------------

C:\Python34\lib\site-packages\pandas\core\reshape.py in unstack(obj, level)
    439     if isinstance(obj, DataFrame):
    440         if isinstance(obj.index, MultiIndex):
--> 441             return _unstack_frame(obj, level)
    442         else:
    443             return obj.T.stack(dropna=False)

C:\Python34\lib\site-packages\pandas\core\reshape.py in _unstack_frame(obj, level)
    479     else:
    480         unstacker = _Unstacker(obj.values, obj.index, level=level,
--> 481                                value_columns=obj.columns)
    482         return unstacker.get_result()
    483

C:\Python34\lib\site-packages\pandas\core\reshape.py in __init__(self, values, index, level, value_columns)
    101
    102         self._make_sorted_values_labels()
--> 103         self._make_selectors()
    104
    105     def _make_sorted_values_labels(self):

C:\Python34\lib\site-packages\pandas\core\reshape.py in _make_selectors(self)
    143
    144         if mask.sum() < len(self.index):
--> 145             raise ValueError('Index contains duplicate entries, '
    146                              'cannot reshape')
    147

ValueError: Index contains duplicate entries, cannot reshape
```