@jorisvandenbossche 2015-02-03
@seth-p maybe don't rush with changing it `None`, it was just an idea of me, I don't know what others think of it

And on `unstack`, what does it have to do with the issue you are solving here? Why does this need to be changed?