@seth-p 2015-02-03
Re: ``unstack()``. I would think that everything that applies to ``stack()`` should apply to ``unstack`` -- e.g. passing ``level=None`` to (un)stack all index levels; and passing ``sequentially=True/False`` to specify whether an array of levels should be (un)stacked sequentially or simultaneously -- no? Since ``.unstack()`` is roughly (up to edge cases) equivalent to ``.T.stack().T``, I would think so. I could leave ``unstack()`` unchanged, but it seems deficient relative to ``stack()``.