@jorisvandenbossche 2015-02-03
Ah, yes, when introducing new behaviour/arguments, that should indeed best in both stack/unstack if that makes sense