@jorisvandenbossche 2015-02-03
But how does unstacking all levels look like?