@seth-p 2015-02-03
Well, if ``df.index`` is a ``MultiIndex``, then I would expect/want ``df.unstack(level=None)`` to be the same as ``df.T.stack(level=None).T`` -- i.e. result in a ``Series`` whose index is a ``MultiIndex`` whose levels are those of ``df.columns`` followed by those of ``df.index``. (In comparison, ``df.stack(level=None)`` will result in a ``Series`` whose index is a ``MultiIndex`` whose levels are those of ``df.index`` followed by those of ``df.columns``.)