@jorisvandenbossche 2015-02-03
The behaviour you describe is what `df.unstack()` already does when you have a single index, I think.

```
In [39]: index = pd.MultiIndex.from_tuples([('one', 'a'), ('one', 'b'),
   ....: ...                                    ('two', 'a'), ('two', 'b')])

In [40]: s = pd.Series(np.arange(1.0, 5.0), index=index)

In [42]: s
Out[42]:
one  a    1
     b    2
two  a    3
     b    4
dtype: float64

In [43]: s.unstack()
Out[43]:
     a  b
one  1  2
two  3  4

In [44]: df = s.unstack()

In [45]: df.unstack()
Out[45]:
a  one    1
   two    3
b  one    2
   two    4
dtype: float64
```