@seth-p 2015-02-03
Yes, precisely. But one may want to unstack all levels of the ``.df.index`` simultaneously -- particularly with ``sequentially=False``. I know personally I've found myself wanting to ``stack()`` all column levels simultaneously, and while I haven't had personal need for ``unstack()``, I must think that someone would want it to support that behavior.