@seth-p 2015-02-04
I added update the ``stack()`` code to handle ``NaN`` level values. Since ``unstack()`` already handles them (#9061 and #9292), I believe this closes #9406.