@behzadnouri 2015-02-13
Invoked with :
    --ncalls: 3
    --repeats: 12


    -------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------
    reshape_stack_simple                         |  15.0343 |   5.3796 |   2.7947 |
    reshape_unstack_simple                       | 4494.4494 |   5.0020 | 898.5265 |
    unstack_sparse_keyspace                      | 6584.2943 |   2.4393 | 2699.2114 |
    -------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------

    Ratio < 1.0 means the target commit is faster then the baseline.
    Seed used: 1234

    Target [86a6e51] : ENH/API: DataFrame.stack() supports level=None, sequentially=True/False, and NaN level values.
    Base   [e266c3d] : Merge pull request #9331 from shoyer/categorical-unique-order

    BUG: don't sort unique values from categoricals